pyproject.toml
[project]
name = "e-com"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "booktype>=1.5",
    "fastapi>=0.132.1",
    "greenlet>=3.3.2",
    "jinja2>=3.1.6",
    "orjson>=3.11.7",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
